Improved prepare_rebase script

+++ prepare_rebase.sh
@@ -1,10 +1,17 @@
 #!/bin/sh
 
-set -x
+if [ -z "$1" ]; then
+    echo "Missing target branch: prepare_rebase <target_branch>"
+    exit 1
+fi
+
+target_branch="$1"
+
+set -ex
 
 git reset --hard HEAD
 git checkout develop
 git pull
-git branch -d "$1"
-git checkout "$1"
+git branch -d "$target_branch" || true 
+git checkout "$target_branch"
 git pull
